made row reduction 1 easier to read
